[FIX] Fix constraint raised on package height
The packaging is set before the height is entered in the reception
screen. When the packaging has a height specified it could be ok but
when there is none and the height is required the constraint will be
raised.
Also the order in which the value are set is important.